ui-api/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
elasticsearch==8.17.2
cassandra-driver==3.29.2
